refactor: add proper v2 folder for keyring-api
