+++ packages/keyring-api/src/v2/index.ts
@@ -1,1 +1,1 @@
-export * from '../api/v2';
+export * from './api';
